Added benchmark: searches take ~70ns, baseline against ~55ns in rust standard lib

+++ src/main.rs
@@ -3,7 +3,7 @@ mod search_tree;
 use search_tree::SearchTree;
 
 fn main() {
-    let array: Vec<i32> = (0..8).collect();
+    let array: Vec<i64> = (0..8).collect();
     if let Ok(search_tree) = SearchTree::new(array.into_iter(), 8) {
         search_tree.search(6);
     }
